delete auth done i think

+++ posts/route.js
@@ -43,19 +43,23 @@ router.post('/:userId', authenticate, async (req, res) => {
             return res.status(500).json({ message: 'Server error, please try again later' });
         }     
     } else {
-        return res.status.json({ message: 'Unauthorized.' });
+        return res.status(403).json({ message: 'Unauthorized.' });
     }
 })
 
-router.delete('/:postId', async (req, res) => {
+router.delete('/:postId', authenticate, async (req, res) => {
     const { postId } = req.params;
-    
+    const { userId } = req.user;
+    const postInfo = await Posts.getPostById(postId);
+
+    if(!postInfo) return res.status(404).json({ message: 'Post not found.' })
+    if(userId !== postInfo.userId) return res.status(403).json({ message: 'Unauthorized.' })
+
     try {
         const deletedPost = await Posts.deletePost(postId);
-        return res.status(200).json('deleted')
+        return res.status(200).json(deletedPost);
     } catch {
-        return res.status(500).json('no')
-
+        return res.status(500).json({ message: 'The server says no, try again later.' });
     }
 })
